requirements.txt
flask==3.0.0
twilio==8.10.0
openai>=1.40.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
flask-cors==4.0.0
httpx>=0.27.0
huggingface-hub>=0.20.0